Extend tests for the IngestVersionManagerDao
